CP-6096: Renamed ISCSI_base to a better BaseISCSI
