+++ Makefile
@@ -20,7 +20,7 @@ SM_DRIVERS += SMB
 SM_LIBS := SR
 SM_LIBS += SRCommand
 SM_LIBS += VDI
-SM_LIBS += ISCSI_base
+SM_LIBS += BaseISCSI
 SM_LIBS += cleanup
 SM_LIBS += lvutil
 SM_LIBS += lvmcache
